Fixes for decimal casting and ignore errors

+++ src/execution/operator/csv_scanner/sniffer/type_detection.cpp
@@ -413,7 +413,8 @@ void CSVSniffer::DetectTypes() {
 
 		// it's good if the dialect creates more non-varchar columns, but only if we sacrifice < 30% of
 		// best_num_cols.
-		if (varchar_cols < min_varchar_cols && info_sql_types_candidates.size() > (max_columns_found * 0.7) &&
+		if (varchar_cols<min_varchar_cols &&static_cast<double>(info_sql_types_candidates.size())>(max_columns_found *
+		                                                                                           0.7) &&
 		    (!options.ignore_errors.GetValue() || candidate->error_handler->errors.size() < min_errors)) {
 			min_errors = candidate->error_handler->errors.size();
 			best_header_row.clear();
